+++ src/integrations-react-hook-form.ts
@@ -1,3 +1,2 @@
-export { default as ZDrop } from "./components/ZDrop";
 export { ZDropField } from "./components/ZDrop/integrations/react-hook-form/ZDropField";
 export type * from "./components/ZDrop/types/zDropTypes";