+++ src/stories/types/visualComponentsTypes.ts
@@ -1,2 +1,7 @@
 export type SelectType = "survivors" | "weapons" | "zombies";
-export type StoryType = "numbers" | "strings" | "objects" | "basic";
+export type StoryType =
+  | "numbers"
+  | "strings"
+  | "objects"
+  | "basic"
+  | "integrations";